Move back ParquetSchemaProducer to ParquetVariantUtil
